vlm-fridge-tracker/requirements.txt
opencv-python>=4.9.0
numpy>=1.26.0
google-generativeai>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
sqlmodel>=0.0.16
